+++ src/main/java/Factory/UserFactory.java
@@ -4,7 +4,7 @@ import Domain.User;
 
 public class UserFactory
 {
-    public static User getUser(String name, String surname, String studentNumber, long telNumber)
+    public static User getUser(String name, String surname, String studentNumber, int telNumber)
     {
         return new User.Builder()
                 .name(name)
